fix upload + thumbnail logic

+++ server/src/models/Book.js
@@ -58,9 +58,8 @@ class Book {
             parameters.splice(3, 0, parseInt(genre_id));
         }
 
-        const query = `SELECT books.*, CONCAT('${basePath}', book_images.image_name) AS primary_image_path
+        const query = `SELECT *, CONCAT('${basePath}', thumbnail) AS thumbnail_path
                        FROM books 
-                       LEFT JOIN book_images ON books.id = book_images.book_id AND book_images.is_primary = 1
                        WHERE (books.title LIKE ? OR books.author LIKE ? or books.description LIKE ?) ${genreQuery} AND books.deleted_at IS NULL 
                        ${orderBy} LIMIT ? OFFSET ?`;
 
@@ -69,11 +68,11 @@ class Book {
     }
 
     static getBookById(id) {
-        const bookQuery = 'SELECT * FROM books WHERE id = ? AND deleted_at IS NULL';
-
         const basePath = '/images/';
 
-        const imagesQuery = `SELECT CONCAT('${basePath}', image_name) AS image_path, is_primary FROM book_images WHERE book_id = ?`;
+        const bookQuery = `SELECT *, CONCAT('${basePath}', thumbnail) AS thumbnail_path FROM books WHERE id = ? AND deleted_at IS NULL`;
+
+        const imagesQuery = `SELECT CONCAT('${basePath}', image) AS image_path FROM book_images WHERE book_id = ?`;
 
         return Promise.all([
             this.queryDatabase(bookQuery, [id]),
@@ -89,18 +88,15 @@ class Book {
         });
     }
 
-    static createBook(book, primaryImageName, otherImageNames) {
+    static createBook(book, otherImages) {
         return this.queryDatabase('INSERT INTO books SET ?', book) // specific syntax for inserting data into a table by passing an object where the keys are the column names and the values are the values to be inserted
             .then(results => {
                 const bookId = results.insertId;
                 const imageQueries = [];
 
-                // Insert primary image
-                imageQueries.push(this.queryDatabase('INSERT INTO book_images (book_id, image_name, is_primary) VALUES (?, ?, 1)', [bookId, primaryImageName]));
-
                 // Insert other images
-                otherImageNames.forEach(imageName => {
-                    imageQueries.push(this.queryDatabase('INSERT INTO book_images (book_id, image_name) VALUES (?, ?)', [bookId, imageName]));
+                otherImages.forEach(image => {
+                    imageQueries.push(this.queryDatabase('INSERT INTO book_images (book_id, image) VALUES (?, ?)', [bookId, image]));
                 });
 
                 return Promise.all(imageQueries)
@@ -108,17 +104,18 @@ class Book {
             });
     }
 
-    static updateBook(id, book) {
+    static updateBook(id, book, otherImages) {
         return this.queryDatabase('UPDATE books SET ? WHERE id = ?', [book, id])
             .then(() => {
                 const imageQueries = [];
 
-                imageQueries.push(this.queryDatabase('DELETE FROM book_images WHERE book_id = ?', id));
+                if (otherImages && otherImages.length > 0) { // empty array is truthy, so need to check length
+                    imageQueries.push(this.queryDatabase('DELETE FROM book_images WHERE book_id = ?', id));
 
-                imageQueries.push(this.queryDatabase('INSERT INTO book_images (book_id, image_name, is_primary) VALUES (?, ?, 1)', [id, book.primaryImage]));
-
-                imageQueries.push(this.queryDatabase('INSERT INTO book_images (book_id, image_name) VALUES (?, ?)', [id, book.otherImages]));
-
+                    otherImages.forEach(image => {
+                        imageQueries.push(this.queryDatabase('INSERT INTO book_images (book_id, image) VALUES (?, ?)', [id, image]));
+                    });
+                }
                 return Promise.all(imageQueries)
                     .then(() => id);
             });
